// src/app/utility/status_flow/enums/inward-status.enum.ts
export enum InwardStatus {
  NOT_STARTED = "NOT_STARTED",
  INWARD_REGISTERED = "INWARD_REGISTERED",
  INWARD_COMPLETED = "INWARD_COMPLETED",
  UNDER_PLANNING = "UNDER_PLANNING",
  UNDER_REVIEW = "UNDER_REVIEW",
  REVIEW_COMPLETED = "REVIEW_COMPLETED",
  IN_PROGRESS = "IN_PROGRESS",
  PARTIALLY_COMPLETED = "PARTIALLY_COMPLETED",
  COMPLETED = "COMPLETED",
  REJECTED = "REJECTED",
  CANCELLED = "CANCELLED"
}